requirements.txt
# Core FastAPI Framework
fastapi
uvicorn[standard]

# Database & ORM
sqlalchemy
psycopg2-binary
alembic

# Authentication & Security
python-jose
passlib
python-multipart
bcrypt

# Configuration & Environment
pydantic
pydantic-settings
python-dotenv

# Testing
pytest
pytest-asyncio
httpx

# Logging & Monitoring
loguru
sentry-sdk

# Developer Tools
black
mypy
ruff
ipython

# Optional Performance
orjson
